Custom doze: use ExecutorService to handle sensor listener
maybe this is useless, but out of curiosity let's try what Google
did for systemui here:
https://github.com/ezio84/abc_frameworks_base/commit/fabc743bcf6e6623e530545c4b31285ea642f087

registering a sensor seems to be an expensive operation, and we do it on each
screen-on event, so moving it to an asynchronous task looks like
a good idea anyway.

The interesting part about ExecutorService api:
By
moving all non-essential binder calls of the main thread or to the
next frame, we bring this down to 5ms, such that window animation
and Keyguard animation starts about at the same time.
https://developer.android.com/reference/java/util/concurrent/ExecutorService

Change-Id: Idea5b7937303897abb9484c07536ddcc181dc38f
Signed-off-by: Jagrav Naik <jagravnaik0@gmail.com>